Move animator stuff to motion-kit and update examples
