+++ examples/build.gradle
@@ -27,6 +27,12 @@ android {
     }
 }
 
+repositories {
+    flatDir {
+        dirs '/Users/lintonye/git-repos/motion-kit/library/build/outputs/aar'
+    }
+}
+
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile 'com.android.support:appcompat-v7:21.0.3'
@@ -35,4 +41,7 @@ dependencies {
 
     // compile 'com.jimulabs.mirrorsandbox:mirror-sandbox:0.1.+'
     compile project(':lib')
+
+    //TODO change to maven central reference
+    compile 'com.jimulabs.motionkit:library-debug:0.1.0@aar'
 }
