+++ examples/src/main/java/com/jimulabs/googlemusicmock/box/ChartSandbox.java
@@ -6,8 +6,7 @@ import android.view.View;
 
 import com.jimulabs.googlemusicmock.ChartView;
 import com.jimulabs.googlemusicmock.R;
-import com.jimulabs.mirrorsandbox.MirrorAnimator;
-import com.jimulabs.mirrorsandbox.MirrorAnimatorSandbox;
+import com.jimulabs.motionkit.MirrorAnimator;
 
 import java.util.ArrayList;
 import java.util.Collections;
