+++ examples/src/main/java/com/jimulabs/googlemusicmock/box/MirrorAnimatorSandbox.java
@@ -1,7 +1,14 @@
-package com.jimulabs.mirrorsandbox;
+package com.jimulabs.googlemusicmock.box;
 
 import android.content.Context;
 import android.view.View;
+
+import com.jimulabs.mirrorsandbox.MirrorSandbox;
+import com.jimulabs.mirrorsandbox.MirrorSandboxBase;
+import com.jimulabs.motionkit.MirrorAnimator;
+import com.jimulabs.motionkit.MirrorView;
+import com.jimulabs.motionkit.MirrorWrap;
+import com.jimulabs.motionkit.MotionKit;
 
 import java.util.List;
 
@@ -17,13 +24,11 @@ import java.util.List;
  *
  * Created by lintonye on 14-12-16.
  */
-public abstract class MirrorAnimatorSandbox implements MirrorSandbox {
+public abstract class MirrorAnimatorSandbox extends MirrorSandboxBase {
     private static final String LOG_TAG = "MirrorAnimatorSandbox";
 
-    private final View mRootView;
-
     public MirrorAnimatorSandbox(View root) {
-        mRootView = root;
+        super(root);
         setGlobalSpeed(1);
     }
 
@@ -59,11 +64,11 @@ public abstract class MirrorAnimatorSandbox implements MirrorSandbox {
     }
 
     protected MirrorAnimator together(MirrorAnimator... animators) {
-        return AnimatorUtils.together(getContext(), animators);
+        return MotionKit.together(getContext(), animators);
     }
 
     protected MirrorAnimator together(List<MirrorAnimator> animators) {
-        return AnimatorUtils.together(getContext(), animators);
+        return MotionKit.together(getContext(), animators);
     }
 
     protected MirrorAnimator sq(MirrorAnimator... animators) {
@@ -71,7 +76,7 @@ public abstract class MirrorAnimatorSandbox implements MirrorSandbox {
     }
 
     protected MirrorAnimator sq(List<MirrorAnimator> animators) {
-        return AnimatorUtils.sequence(getContext(), animators);
+        return MotionKit.sequence(getContext(), animators);
     }
 
     protected MirrorWrap wrapToAnimate(Object obj) {
@@ -79,14 +84,14 @@ public abstract class MirrorAnimatorSandbox implements MirrorSandbox {
     }
 
     protected MirrorAnimator sequence(MirrorAnimator... animators) {
-        return AnimatorUtils.sequence(getContext(), animators);
+        return MotionKit.sequence(getContext(), animators);
     }
 
     protected MirrorAnimator sequence(List<MirrorAnimator> animators) {
-        return AnimatorUtils.sequence(getContext(), animators);
+        return MotionKit.sequence(getContext(), animators);
     }
 
     public void setGlobalSpeed(double speed) {
-        AnimatorUtils.setGlobalSpeed(speed);
+        MotionKit.setGlobalSpeed(speed);
     }
 }
